django/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==5.0.2
django-allauth==0.61.1
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-treebeard==4.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environs==10.3.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
marshmallow==3.21.0
moviepy==1.0.3
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
proglog==0.1.10
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.1.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
wheel==0.42.0
